requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-jose
kafka-python
authlib
python-telegram-bot==13.15
passlib[bcrypt]
httpx
pydantic-settings
itsdangerous
email_validator